First production build—now more debugging...

+++ api/plans/plan.model.js
@@ -6,7 +6,7 @@ function model(sequelize) {
     const attributes = {
         accountId: { type: DataTypes.INTEGER, allowNull: false },
         categoryId: { type: DataTypes.INTEGER, allowNull: false },
-        planAmount: { type: DataTypes.FLOAT(2), defaultValue: '0.00', allowNull: false },
+        planAmount: { type: DataTypes.DECIMAL(12,2), defaultValue: '0.00', allowNull: false },
         created: { type: DataTypes.DATE, allowNull: false, defaultValue: DataTypes.NOW },
         updated: { type: DataTypes.DATE },
     };
